+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## $v
+
+- Enhancing configurable filter tweaks.
+
 ## v170221.4504
 
 - Adding compatibility with multiple editor instances in WP.